+++ src/components/pathpanel/PathAqiBar.js
@@ -1,6 +1,6 @@
 import React from 'react'
 import styled, { css } from 'styled-components'
-import { aqiColors } from '../../../constants'
+import { aqiColors } from '../../constants'
 
 const StyledAqiBar = styled.div`
   display: flex;
